Moved position and piece classes to their own modules
